+++ lib/lightning/collaboration/workflow_reconciler.ex
@@ -171,7 +171,7 @@ defmodule Lightning.Collaboration.WorkflowReconciler do
          _workflow
        ) do
     data =
-      ~w(id cron_expression enabled has_auth_method type)a |> pluck_fields(cs)
+      ~w(id cron_expression enabled type)a |> pluck_fields(cs)
 
     {:insert, Doc.get_array(doc, "triggers"), data}
   end
@@ -182,7 +182,7 @@ defmodule Lightning.Collaboration.WorkflowReconciler do
          _workflow
        ) do
     changes =
-      ~w(cron_expression enabled has_auth_method type)a |> pluck_fields(cs)
+      ~w(cron_expression enabled type)a |> pluck_fields(cs)
 
     item =
       Doc.get_array(doc, "triggers") |> find_in_array(cs.data.id)